Move Hash Algorithms to that namespace under Hashing
