+++ tests/Hashing/HashAlgorithms/HashAlgorithmImplementationTest.php
@@ -1,15 +1,15 @@
 <?php
 declare( strict_types = 1 );
 
-namespace PHP\Tests\Hashing;
+namespace PHP\Tests\Hashing\HashAlgorithms;
 
 use PHP\Collections\ByteArray;
-use PHP\Hashing\IHashAlgorithm;
-use PHP\Hashing\MD5;
-use PHP\Hashing\SHA1;
-use PHP\Hashing\SHA256;
-use PHP\Hashing\SHA384;
-use PHP\Hashing\SHA512;
+use PHP\Hashing\HashAlgorithms\IHashAlgorithm;
+use PHP\Hashing\HashAlgorithms\MD5;
+use PHP\Hashing\HashAlgorithms\SHA1;
+use PHP\Hashing\HashAlgorithms\SHA256;
+use PHP\Hashing\HashAlgorithms\SHA384;
+use PHP\Hashing\HashAlgorithms\SHA512;
 use PHPUnit\Framework\TestCase;
 
 /**
